Rewrite Redux in Duck style #17
